Added mpool support for matrix module
Matrix module now uses the mpool library for the matrix_t type. All
allocations are now done within init_matrix() and they use the mpool
interface for doing so.

By doing this, the amount of allocations done in a very basic program
(xor_test as a benchmark) is vastly reduced. This should also lead
to minimally faster execution times as malloc isn't being called
millions of times.

Note that fully deallocating the pool currently relies on using
__attribute__ ((destructor)) to call free_mpool() after the program
ends. In the future this will change to have the net "own" it's own
memory pool, and the pool will exist for as long as the net does.